Breaking Reattribution/Redesigntation into own Transaction type
